change package name from laravel-auth to lockminds-auth to avoid default behavour of removing laravel prefix

+++ src/Traits/CustomClaimsAccessTokenTrait.php
@@ -20,9 +20,8 @@ trait CustomClaimsAccessTokenTrait
             ->where('status', 'valid')
             ->first();
 
-        if ($otp->status == 'valid') {
+        if (!empty($otp) && $otp->status == 'valid') {
             \Log::info(' valid otp found');
-
             Otp::where('id', $this->getUserIdentifier())->delete();
         } else {
             \Log::info('No valid otp found');
